Update auf SVN 1.67, Teil 4 mit Änderungen der Arduino-Libraries

+++ libraries/ttydata/ttydata.cpp
@@ -32,7 +32,7 @@ uint8_t TTYdataClass::callfn(char *buf){
 void TTYdataClass::analyze_ttydata(uint8_t channel)
 {
   static char cmdbuf[TTY_BUFSIZE+1];
-  static uint8_t cmdlen;
+  static int cmdlen;
   uint8_t ucCommand;
   uint8_t odc = display.channel;
